core/src/main/java/org/teste/calebe/gerenciamentodetarefascore/service/MyService.java
package org.teste.calebe.gerenciamentodetarefascore.service;

import org.springframework.boot.context.properties.EnableConfigurationProperties;
import org.springframework.stereotype.Service;

@Service
@EnableConfigurationProperties(ServiceProperties.class)
public class MyService {
    private final ServiceProperties serviceProperties;

    public MyService(ServiceProperties serviceProperties) {
        this.serviceProperties = serviceProperties;
    }

    public String message() {
        serviceProperties.getClass();
        return this.serviceProperties.getMessage();
    }
}
